Add formatting of DOS4-7, but not tested. Update hst amiga with formatting fixes. Fix rdb init disk size calculation.

+++ src/Hst.Imager.Core/Commands/RdbInitCommand.cs
@@ -56,7 +56,7 @@
 
                 diskGeometry = new DiskGeometry
                 {
-                    DiskSize = cylinders * heads * sectors * 512,
+                    DiskSize = (long)cylinders * heads * sectors * 512,
                     Cylinders = cylinders,
                     Heads = heads,
                     Sectors = sectors
